backend/pyproject.toml
[project]
name = "our-table-backend"
version = "0.1.0"
description = "Our Table recipe manager backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "recipe-scrapers>=15.0.0",
    "authlib>=1.4.0",
    "python-jose[cryptography]>=3.4.0",
    "httpx>=0.28.0",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.20",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "respx>=0.22.0",
    "pytest-cov>=6.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "respx>=0.22.0",
    "pytest-cov>=6.0.0",
]
